% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fit.R
\name{compare_fit}
\alias{compare_fit}
\title{FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis}
\usage{
compare_fit(model_1, model_2, data)
}
\arguments{
\item{model_1}{1st model input for comparison fit}

\item{model_2}{2nd model input for comparison fit}
}
\description{
FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis
}
\examples{
: compare_fit(model1, model2)
}
