% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_FACT_list.R
\name{create_FACT_list}
\alias{create_FACT_list}
\title{Create a FACT file template}
\usage{
create_FACT_list(
  title = NA,
  source = NA,
  author_orcid = NA,
  reporter = "Wojciech Francuzik",
  description = "",
  implication = "",
  data = NA,
  vis = F,
  stat_test = NA
)
}
\arguments{
\item{title}{The tile for the fact that you want to desctibe}

\item{source}{The source for the fact that you are describing. It is either DOI or another fact file or an internet link}

\item{author_orcid}{ORCID of the author of the fact that you are describing}

\item{reporter}{Usually the person who is reporting this - you}

\item{description}{The description of the fact that you are reporting - with links to relevant research}

\item{implication}{The implications and hypotheses that can be drawn from this fact}

\item{data}{A data frame}

\item{vis}{Boolean if a visaualisation should be present}

\item{stat_test}{result of a statistical test used}
}
\value{
A FACT list
}
\description{
Create a FACT file template
}
