% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factOR.R
\docType{package}
\name{factOR}
\alias{factOR}
\title{factOR Package}
\description{
This package contains functions to facilitate the principles of Fact Oriented Research.
}
\author{
Wojciech Francuzik \email{wolass@gmail.com}
}
