% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get_alpha}
\alias{get_alpha}
\title{FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis}
\usage{
get_alpha(data, n, group_name)
}
\arguments{
\item{data}{the dataset you would like to conduct analysis on, must be stored as a .csv file}

\item{n}{the number of factors desired for Exploratory Factor Analysis (EFA); standard value for this is 1, but customize as needed}

\item{group_name}{title you would like to give to the dataset that you are analyzing, must be entered as a character string, in quotation marks (i.e. "Normal Hearing Data")}

\description{
FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis
}
\examples{
: get_alpha(NH_data, 1, "Normal Hearing Data")
}
\keyword{alpha}
\keyword{analysis,}
\keyword{cronbach's}
\keyword{factor}
