% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getloadingsmodel.R
\name{get_loadings_model}
\alias{get_loadings_model}
\title{FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis}
\usage{
get_loadings_model(model_input, data_input, group_name)
}
\arguments{
\item{model_input}{the model used for comparison in Confirmatory Factor Analysis (CFA)}

\item{data_input}{the dataset you would like to conduct analysis on (like the 'data' param), must be stored as a .csv file}

\item{group_name}{title you would like to give to the dataset that you are analyzing, must be entered as a character string, in quotation marks (i.e. "Normal Hearing Data")}
}
\description{
FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis
}
\examples{
: get_loadings_model(model, NH_data, "Normal Hearing Data")
}
\keyword{alpha}
\keyword{analysis,}
\keyword{cronbach's}
\keyword{factor}
