% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handy_fit.R
\name{handy_fit}
\alias{handy_fit}
\title{FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis}
\usage{
handy_fit(model_input, data_input)
}
\arguments{
\item{model_input}{the model used for comparison in Confirmatory Factor Analysis (CFA)}

\item{data_input}{the dataset you would like to conduct analysis on (like the 'data' param), must be stored as a .csv file}
}
\description{
FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis
}
\examples{
: handy_fit(model, NH_data)
}
\keyword{alpha,}
\keyword{analysis,}
\keyword{cronbach's}
\keyword{factor}
\keyword{fit}
\keyword{handy}
