% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_check.R
\name{pick_check}
\alias{pick_check}
\title{FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis}
\usage{
pick_check(test, data)
}
\arguments{
\item{test}{which test you would like to use for determining whether a dataset is suitable for factor analysis, must be entered as a character string inside quotation marks (i.e. "KMO", "Kaiser, Meyer, Olkin", "Kaiser", "Meyer", "Olkin")}

\item{data}{the dataset you would like to conduct analysis on, must be stored as a .csv file}
}
\description{
FACTOR- a complete package for Exploratory and Confirmatory Factor Analysis
}
\examples{
: get_loadings("KMO", NH_data)
}
\keyword{alpha,}
\keyword{analysis,}
\keyword{check}
\keyword{cronbach's}
\keyword{factor}
\keyword{pick}
