% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACTOR-package.R
\docType{package}
\name{FACTOR-package}
\alias{FACTOR}
\alias{FACTOR-package}
\title{FACTOR: Package}
\description{
The function included in this package lets you explore factor analysis. The package contain the function for running both exploratory and confirmatory factor analysis. For confirmatory factor analysis, it includes the code for obtaining factor loadings for all the individual items,cronbach's alpha value, and statistics for model fit. Whereas, for exploratory factor analysis, it contain function for obtaining factor laodings and cronbach's alpha value.
}
\author{
\strong{Maintainer}: Sajana Aryal \email{sajanaaryal5566@gmail.com}

Authors:
\itemize{
  \item Yao Chen \email{chenyao@utexas.edu}
  \item Keerthi Stanley \email{keerthi.stanley@utexas.edu}
}

}
\keyword{internal}
